+++ CHANGELOG.md
@@ -1,6 +1,28 @@
-## 0.1.0 (Aug 8, 2022)
+# Changelog
 
-FEATURES
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+
+## [Unreleased]
+
+### Added
+
+- **New Resource** `ns_device_profile` ([#5])
+- Migrated the provider to terraform-plugin-framework.
+- Updated the resource name from `citrixsdx` to `netscalersdx`.
+
+
+## [0.1.0] - 2022-08-08
+
+### Added
 
 * **New Resource** `provision_vpx`
 * **New Resource** `vpx_state`
+
+
+[Unreleased]: https://github.com/netscaler/terraform-provider-citrixsdx/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/netscaler/terraform-provider-citrixsdx/releases/tag/v0.1.0
+[#5]: https://github.com/netscaler/terraform-provider-citrixsdx/issues/5